blofin-stack/requirements.txt
websockets==12.0
python-dotenv==1.0.1
ccxt==4.4.31
